+++ apps/dev-launchpad/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @stackframe/dev-launchpad
 
+## 2.8.56
+
+## 2.8.55
+
+## 2.8.54
+
+### Patch Changes
+
+- Various changes
+
 ## 2.8.53
 
 ## 2.8.52